Add where param to support conditional reads
The optional where parameter is passed through as part of the mongo query to retrieve an entity. For example, this
lets the caller fetch an entity only if it's timestamp is more recent than the one provided by the caller. The query
includes a match on the entity id 'anded' with the where clause so an empty result could be because either the
where clause failed or there was no match on the entity id.