Switch to Parse service for push notifications: ** API change **
We now use Parse for all notifications and can support the same platforms that
Parse does. See registerInstall method for details on the data we need from clients.
Android clients will not break with this change but they will not receive any
notifications until they start sending the parseInstallId instead of the deprecated
registrationId